@@ -10,11 +10,13 @@
   },
   "author": "",
   "license": "ISC",
-  "dependencies": {
-    "lodash": "^4.17.21"
-  },
   "devDependencies": {
-    "eslint": "^8.11.0",
+    "@swc/core": "^1.3.4",
+    "@swc/jest": "^0.2.23",
+    "@types/jest": "^29.1.1",
+    "eslint": "^8.24.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.26.0",
     "jest": "^27.5.1"
   }
 }
