@@ -1,22 +1,20 @@
 module.exports = {
   env: {
-    es6: true,
+    commonjs: true,
+    es2021: true,
     node: true,
     jest: true,
   },
   extends: [
-    'eslint:recommended',
+    'airbnb-base',
   ],
   parserOptions: {
-    'ecmaVersion': 2016,
+    ecmaVersion: 'latest',
   },
   rules: {
-    'indent': ['error', 2],
-    'linebreak-style': ['error', 'unix'],
-    'quotes': ['error', 'single'],
-    'semi': ['error', 'always'],
+    indent: ['error', 2],
     'no-trailing-spaces': 'error',
-    'curly': 'error',
+    curly: 'error',
     'brace-style': 'error',
     'no-multi-spaces': 'error',
     'space-infix-ops': 'error',
@@ -24,15 +22,15 @@ module.exports = {
     'no-whitespace-before-property': 'error',
     'func-call-spacing': 'error',
     'space-before-blocks': 'error',
-    'keyword-spacing': ['error', { 'before': true, 'after': true }],
-    'comma-spacing': ['error', { 'before': false, 'after': true }],
+    'keyword-spacing': ['error', { before: true, after: true }],
+    'comma-spacing': ['error', { before: false, after: true }],
     'comma-style': ['error', 'last'],
     'comma-dangle': ['error', 'always-multiline'],
     'space-in-parens': ['error', 'never'],
     'block-spacing': 'error',
     'array-bracket-spacing': ['error', 'never'],
     'object-curly-spacing': ['error', 'always'],
-    'key-spacing': ['error', { 'mode': 'strict' }],
-    'arrow-spacing': ['error', { 'before': true, 'after': true }],
+    'key-spacing': ['error', { mode: 'strict' }],
+    'arrow-spacing': ['error', { before: true, after: true }],
   },
 };
